requirements.txt
firecrawl-py
pyyaml 